Move sync packet I/O out of library and into tests/client
read_exact, read_packet, and write_packet were only ever used by the
test client and test helpers, but lived in src/hegel/protocol/packet.py
with a comment noting this. Move them to tests/client/protocol.py
where they belong.

Update test_packet.py and test_connection.py to import from their
new location. Remove import socket from packet.py (only needed for
the sync API).
